Add TrakDiag V2.4.1

+++ Logical/Trak/Libraries/TrakDiag/TD_Recorder.cpp
@@ -488,8 +488,8 @@ void TD_Recorder(struct TD_Recorder* inst)
 							}
 						}
 						if( indexOfSegment >= 0 ){ 
-							int len = sprintf( temp, "%s new Shuttle( %d, %d, 0x%x, %i, %lf )", !validEntriesFound ? "" : "\n\t\t\t\t,", 
-																			pEntry->Index,
+							int len = sprintf( temp, "%s new Shuttle( %lu, %d, 0x%x, %i, %lf )", !validEntriesFound ? "" : "\n\t\t\t\t,", 
+																			(unsigned long) pEntry->Index,
 																			pEntry->Active,
 																			pEntry->PLCopenState,
 																			indexOfSegment, pEntry->SegmentPosition );
